feat: override UW and HighDelay pose estimators callbacks
This overrides the callbacks for xy_position, orientation and depth
samples. The new versions use a fixed covariance for all these methods.
Without this, it is troublesome to use some pose sources that dont
provide cov_position and cov_orientation, such as those that come from a
transformer and that do some transformations inside of them.